Removed unnecessary cast

+++ cmd/engine/engine.go
@@ -94,7 +94,7 @@ const (
 
 	DefaultAPIDefaultUser               = "clouditor"
 	DefaultAPIDefaultPassword           = "clouditor"
-	DefaultAPIgRPCPort                  = uint16(9090)
+	DefaultAPIgRPCPort                  = 9090
 	DefaultAPIStartEmbeddedOAuth2Server = true
 	DefaultServiceOAuth2Endpoint        = "http://localhost:8080/v1/auth/token"
 	DefaultServiceOAuth2ClientID        = "clouditor"
@@ -103,7 +103,7 @@ const (
 	DefaultDBPassword                   = "postgres"
 	DefaultDBHost                       = "localhost"
 	DefaultDBName                       = "postgres"
-	DefaultDBPort                       = uint16(5432)
+	DefaultDBPort                       = 5432
 	DefaultDBInMemory                   = false
 	DefaultCreateDefaultTarget          = true
 	DefaultDiscoveryAutoStart           = false
